# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pay for Nothing API",
    "version": "1.0.0",
    "description": "Users pay simulated money for nothing and compete on a public leaderboard."
  },
  "servers": [{ "url": "http://localhost:3000/api" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "created_at": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "summary": "Get your profile",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "description": "Not authenticated" }
        }
      },
      "put": {
        "summary": "Update display name",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username"],
                "properties": { "username": { "type": "string", "example": "NewName" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Too short" },
          "409": { "description": "Name taken" }
        }
      },
      "delete": {
        "summary": "Delete account and all donations",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Not authenticated" }
        }
      }
    },
...
  }
}